feat: 비가격 AlignedMarketData 필드 float32 다운캐스트
- **Why:** bridge_post_align의 OHLC 외 9개 비가격 필드(volume/funding/basis/oi/lsr 등)를 float64로 저장하여 12GiB RSS 예산을 초과하였으며, quant.md 정밀도 정책상 OHLC만 float64가 필요하고 나머지는 float32로 충분하였음.
- **What:** AlignedMarketData의 volume_2d·funding_2d 등 9개 필드와 align_data_maps 내 초기화·대입 dtype을 float64→float32로 변경하고, ablation.py aligned_data 타입과 L2SimulationCache hurdle_2d·funding_2d 타입을 NDArray[floating[Any]]로 완화하였음. OHLC 4개 필드는 float64를 유지하고, float32 volume_2d가 _zscore_2d에서 float64 기준선 대비 rtol 1e-3 이내로 일치함을 검증하는 통합 테스트를 추가하였음.

+++ src/domain/futures/strategy/common/alignment.py
@@ -37,19 +37,19 @@ class AlignedMarketData:
     high_2d: NDArray[np.float64]
     low_2d: NDArray[np.float64]
     close_2d: NDArray[np.float64]
-    volume_2d: NDArray[np.float64]
-    funding_2d: NDArray[np.float64]
+    volume_2d: NDArray[np.float32]
+    funding_2d: NDArray[np.float32]
     active_mask: NDArray[np.bool_]
     warm_mask: NDArray[np.bool_]
     entry_block_mask: NDArray[np.bool_]
     kill_mask: NDArray[np.bool_]
-    basis_2d: NDArray[np.float64] | None = None
-    oi_2d: NDArray[np.float64] | None = None
-    lsr_2d: NDArray[np.float64] | None = None
-    taker_buy_2d: NDArray[np.float64] | None = None
-    trades_2d: NDArray[np.float64] | None = None
-    adv_usdt_2d: NDArray[np.float64] | None = None
-    execution_cost_bps_2d: NDArray[np.float64] | None = None
+    basis_2d: NDArray[np.float32] | None = None
+    oi_2d: NDArray[np.float32] | None = None
+    lsr_2d: NDArray[np.float32] | None = None
+    taker_buy_2d: NDArray[np.float32] | None = None
+    trades_2d: NDArray[np.float32] | None = None
+    adv_usdt_2d: NDArray[np.float32] | None = None
+    execution_cost_bps_2d: NDArray[np.float32] | None = None
     # Phase D: C1 inference panel 전용 마스크 (Stage5 timeline 기반). None이면 미사용.
     inference_active_mask: NDArray[np.bool_] | None = None
     inference_entry_warm_mask: NDArray[np.bool_] | None = None
@@ -148,15 +148,15 @@ def align_data_maps(
     high_2d: NDArray[np.float64] = np.zeros((eff_len, n), dtype=np.float64)
     low_2d: NDArray[np.float64] = np.zeros((eff_len, n), dtype=np.float64)
     close_2d: NDArray[np.float64] = np.zeros((eff_len, n), dtype=np.float64)
-    volume_2d: NDArray[np.float64] = np.zeros((eff_len, n), dtype=np.float64)
-    funding_2d: NDArray[np.float64] = np.zeros((eff_len, n), dtype=np.float64)
-    basis_2d: NDArray[np.float64] | None = None
-    oi_2d: NDArray[np.float64] | None = None
-    lsr_2d: NDArray[np.float64] | None = None
-    taker_buy_2d: NDArray[np.float64] | None = None
-    trades_2d: NDArray[np.float64] | None = None
-    adv_usdt_2d: NDArray[np.float64] = np.full((eff_len, n), np.nan, dtype=np.float64)
-    execution_cost_bps_2d: NDArray[np.float64] = np.full((eff_len, n), np.nan, dtype=np.float64)
+    volume_2d: NDArray[np.float32] = np.zeros((eff_len, n), dtype=np.float32)
+    funding_2d: NDArray[np.float32] = np.zeros((eff_len, n), dtype=np.float32)
+    basis_2d: NDArray[np.float32] | None = None
+    oi_2d: NDArray[np.float32] | None = None
+    lsr_2d: NDArray[np.float32] | None = None
+    taker_buy_2d: NDArray[np.float32] | None = None
+    trades_2d: NDArray[np.float32] | None = None
+    adv_usdt_2d: NDArray[np.float32] = np.full((eff_len, n), np.nan, dtype=np.float32)
+    execution_cost_bps_2d: NDArray[np.float32] = np.full((eff_len, n), np.nan, dtype=np.float32)
     active_mask: NDArray[np.bool_] = np.ones((eff_len, n), dtype=bool)
     warm_mask: NDArray[np.bool_] = np.ones((eff_len, n), dtype=bool)
     entry_block_mask: NDArray[np.bool_] = np.zeros((eff_len, n), dtype=bool)
@@ -264,33 +264,33 @@ def align_data_maps(
             funding_2d[:, col] = block.numeric[:, n_idx["funding_rate"]]
         if "basis" in n_idx:
             if basis_2d is None:
-                basis_2d = np.full((eff_len, n), np.nan, dtype=np.float64)
+                basis_2d = np.full((eff_len, n), np.nan, dtype=np.float32)
             basis_2d[:, col] = block.numeric[:, n_idx["basis"]]
         elif "basis_rate" in n_idx:
             if basis_2d is None:
-                basis_2d = np.full((eff_len, n), np.nan, dtype=np.float64)
+                basis_2d = np.full((eff_len, n), np.nan, dtype=np.float32)
             basis_2d[:, col] = block.numeric[:, n_idx["basis_rate"]]
         for _oi_alias in ("sum_open_interest", "open_interest", "oi"):
             if _oi_alias in n_idx:
                 if oi_2d is None:
-                    oi_2d = np.full((eff_len, n), np.nan, dtype=np.float64)
+                    oi_2d = np.full((eff_len, n), np.nan, dtype=np.float32)
                 oi_2d[:, col] = block.numeric[:, n_idx[_oi_alias]]
                 break
         for _lsr_alias in ("long_short_ratio", "global_long_short_ratio"):
             if _lsr_alias in n_idx:
                 if lsr_2d is None:
-                    lsr_2d = np.full((eff_len, n), np.nan, dtype=np.float64)
+                    lsr_2d = np.full((eff_len, n), np.nan, dtype=np.float32)
                 lsr_2d[:, col] = block.numeric[:, n_idx[_lsr_alias]]
                 break
         for _taker_alias in ("taker_buy_base", "taker_buy_quote"):
             if _taker_alias in n_idx:
                 if taker_buy_2d is None:
-                    taker_buy_2d = np.full((eff_len, n), np.nan, dtype=np.float64)
+                    taker_buy_2d = np.full((eff_len, n), np.nan, dtype=np.float32)
                 taker_buy_2d[:, col] = block.numeric[:, n_idx[_taker_alias]]
                 break
         if "trades" in n_idx:
             if trades_2d is None:
-                trades_2d = np.full((eff_len, n), np.nan, dtype=np.float64)
+                trades_2d = np.full((eff_len, n), np.nan, dtype=np.float32)
             trades_2d[:, col] = block.numeric[:, n_idx["trades"]]
         if "adv_usdt" in n_idx:
             adv_usdt_2d[:, col] = block.numeric[:, n_idx["adv_usdt"]]
